requirements.base.txt

pyicloud
python-dotenv
